feat(admins): move to WorldProfileWidget, rename to WorldMembers
