+++ packages/app/src/core/enums/creatorTabs.enum.ts
@@ -1,6 +1,5 @@
 export enum CreatorTabsEnum {
   addObject = 'addObject',
-  admins = 'admins',
   gizmo = 'gizmo',
   skybox = 'skybox',
   sound = 'sound',
