--- packages/app/src/scenes/widgets/pages/CreatorWidget/pages/Admins/index.ts
@@ -1,1 +1,0 @@
-export {default as Admins} from './Admins';
