+++ packages/app/src/scenes/widgets/pages/CreatorWidget/pages/index.ts
@@ -4,4 +4,3 @@ export * from './ObjectInspector';
 export * from './SkyboxSelector';
 export * from './SpawnAsset';
 export * from './MusicManager';
-export * from './Admins';
