+++ packages/app/src/scenes/widgets/pages/WorldProfileWidget/components/WorldMembers/index.ts
@@ -1,0 +1,1 @@
+export {default as WorldMembers} from './WorldMembers';
