+++ packages/app/src/scenes/widgets/pages/WorldProfileWidget/components/index.ts
@@ -1,2 +1,3 @@
 export * from './WorldView';
 export * from './WorldEditor';
+export * from './WorldMembers';
